confidential: Move `isOCPVersionLessThan` to confidential code
There are no other users for this but baremetal CoCo. Also,
the checking logic requires a massive rewrite to support multiple
OCP minor versions. Regroup the code to ease that work.

No behavior change.
