+++ controllers/confidential_handler.go
@@ -5,8 +5,11 @@ import (
 	"fmt"
 	"maps"
 
+	semver "github.com/Masterminds/semver/v3"
+	configv1 "github.com/openshift/api/config/v1"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -153,6 +156,34 @@ func (r *KataConfigOpenShiftReconciler) handleConfidentialPeerPods(state Feature
 	}
 
 	return nil
+}
+
+// isOCPVersionLessThan checks if the current OpenShift cluster version is less than the specified minimum version.
+// Returns (true, version, nil) if current version < minVersion, (false, version, nil) if current >= minVersion.
+// Returns error if cluster version is not available or cannot be retrieved.
+func (r *KataConfigOpenShiftReconciler) isOCPVersionLessThan(minVersion string) (bool, string, error) {
+	clusterVersion := &configv1.ClusterVersion{}
+	err := r.Client.Get(context.TODO(), types.NamespacedName{Name: "version"}, clusterVersion)
+	if err != nil {
+		return false, "", err
+	}
+
+	currentVersion := clusterVersion.Status.Desired.Version
+	if currentVersion == "" {
+		return false, "", fmt.Errorf("cluster version not available yet")
+	}
+
+	current, err := semver.NewVersion(currentVersion)
+	if err != nil {
+		return false, currentVersion, fmt.Errorf("unable to parse current OCP version %s: %w", currentVersion, err)
+	}
+
+	min, err := semver.NewVersion(minVersion)
+	if err != nil {
+		return false, currentVersion, fmt.Errorf("invalid minimum version format %s: %w", minVersion, err)
+	}
+
+	return current.LessThan(min), currentVersion, nil
 }
 
 // handleConfidentialBaremetal configures confidential computing for baremetal deployments.
